+++ changelog.md
@@ -10,3 +10,13 @@
 
 ## Version 2 
 1. added mySql, inquirer, readme, gitignore.
+
+## Version 3
+1. Added schema and seeds sql.
+
+## Version 4
+1. added change shells to 3 .js's.
+
+## Version 5
+1. attempted to write the logic to the 3 js files.
+2. included the npm console.table